Add exports. refector methods into another file. Bump version
