+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "flow-cadut",
-  "version": "0.1.15-alpha.22",
+  "version": "0.1.15-alpha.23",
   "description": "Flow Cadence Template Utilities",
   "author": "Maksim Daunarovich",
   "license": "Apache-2.0",
