+++ src/index.js
@@ -63,6 +63,11 @@ export {
 
 export { PLUGIN_TYPES, registerPlugin, getPlugins } from "./plugins";
 
+export { CURRENT_SPORK_ROOT, CURRENT_SPORK_NUMBER } from "./const";
+export { getEventName, getEventsInRange, findLatestEvents } from "./events";
+
+export { getChainHeight, getLatestBlock } from "./chain";
+
 export { executeScript, sendTransaction, deployContract, updateContract } from "./interactions";
 
 // Below is a set of aliases to bring it in line with FCL
